--- a/crates/telio-model/src/features.rs
+++ b/crates/telio-model/src/features.rs
@@ -3,6 +3,7 @@
 
 use std::{collections::HashSet, fmt};
 
+use ipnet::Ipv4Net;
 use num_enum::{IntoPrimitive, TryFromPrimitive};
 use serde::{de::IntoDeserializer, Deserialize, Deserializer, Serialize};
 use smart_default::SmartDefault;
@@ -353,6 +354,9 @@ pub struct FeatureFirewall {
     /// Turns on connection resets upon VPN server change (Deprecated alias for neptun_reset_conns)
     #[serde(default)]
     pub boringtun_reset_conns: bool,
+    /// Customizable private IP range to treat certain private IP ranges
+    /// as public IPs for testing purposes.
+    pub custom_private_ip_range: Option<Ipv4Net>,
 }
 
 /// Turns on post quantum VPN tunnel
@@ -540,7 +544,8 @@ mod tests {
             "nicknames": true,
             "firewall": {
                 "neptun_reset_conns": true,
-                "boringtun_reset_conns": true
+                "boringtun_reset_conns": true,
+                "custom_private_ip_range": null
             },
             "flush_events_on_stop_timeout_seconds": 15,
             "post_quantum_vpn": {
@@ -630,6 +635,7 @@ mod tests {
                     firewall: FeatureFirewall {
                         neptun_reset_conns: true,
                         boringtun_reset_conns: true,
+                        custom_private_ip_range: None,
                     },
                     flush_events_on_stop_timeout_seconds: Some(15),
                     post_quantum_vpn: FeaturePostQuantumVPN {