--- a/crates/telio-utils/src/lru_cache.rs
+++ b/crates/telio-utils/src/lru_cache.rs
@@ -184,7 +184,7 @@ impl<Key: Clone + Eq + Hash, Value> LruCache<Key, Value> {
 
     /// Gets the given key’s corresponding entry in the map for in-place manipulation.
     #[inline(always)]
-    pub fn entry(&mut self, key: Key) -> Entry<'_, Key, Value> {
+    pub fn entry(&mut self, key: Key, update_last_access_time: bool) -> Entry<'_, Key, Value> {
         let hash = self.map.hasher().hash_one(&key);
         match self.map.raw_entry_mut().from_key_hashed_nocheck(hash, &key) {
             RawEntryMut::Occupied(mut e) => {
@@ -198,7 +198,9 @@ impl<Key: Clone + Eq + Hash, Value> LruCache<Key, Value> {
                         max_map_size: self.capacity,
                     });
                 }
-                Self::update_last_time(&mut e, now);
+                if update_last_access_time {
+                    Self::update_last_time(&mut e, now)
+                }
                 Entry::Occupied(OccupiedEntry {
                     key,
                     hash,
@@ -492,7 +494,7 @@ mod tests {
         assert_eq!(lru_cache.len(), 1);
 
         advance_time_by_ms(300);
-        lru_cache.entry(0);
+        lru_cache.entry(0, true);
         advance_time_by_ms(300);
         assert_eq!(Some(&0), lru_cache.peek(&0));
         advance_time_by_ms(300);
@@ -631,7 +633,7 @@ mod tests {
                             old.remove(&k);
                         },
                         Op::GetUsingEntry(k) => {
-                            let new_val = match new.entry(k) {
+                            let new_val = match new.entry(k, true) {
                                 Entry::Occupied(mut e) => {
                                     Some(*e.get_mut())
                                 },
@@ -642,7 +644,7 @@ mod tests {
                             assert_eq!(new_val, old.get(&k).copied());
                         },
                         Op::GetMutUsingEntry(k) => {
-                            if let Entry::Occupied(mut e) = new.entry(k) {
+                            if let Entry::Occupied(mut e) = new.entry(k, true) {
                                 *e.get_mut() += 1;
                             }
                             if let Some(v) = old.get_mut(&k) {