--- a/nat-lab/tests/test_mesh_api.py
+++ b/nat-lab/tests/test_mesh_api.py
@@ -42,6 +42,8 @@ class TestNode:
             ),
             is_local=False,
             allow_incoming_connections=False,
+            allow_peer_traffic_routing=False,
+            allow_peer_local_network_access=False,
             allow_peer_send_files=False,
             allow_multicast=True,
             peer_allows_multicast=True,