### Problem
Telio-firewall accepted only those connections which had `incoming_connections` as allowed. The problem in it is that if someone just wants to route traffic or allow only local area access, they would also have to allow `incoming_connections`.

### Solution
Update handling incoming packets in firewall, decoupling each permission according to the logic posted in the issue ticket. 

![incoming_connections](https://github.com/NordSecurity/libtelio/assets/50082256/988b7c07-7283-482e-a236-3b202c9ee7fd)

I have divided this PR into 2 parts. First is the main firewall logic, second is complying with all the natlab and unit tests. The boundary may not be very strict but that is the general idea.

The main firewall logic is explained above.
Firewall config was changed to add a custom private ips parameter so that nat-lab tests could access stun server, otherwise it was also being blocked.

Most of the changes in `firewall_bench.rs` is mainly copy paste code in order to comply with new firewall constructor.

The logic in `incoming_tcp_connections` inside `firewall.rs` has been modified to look similar to its udp counterpart to make future maintaining easy as when understands how 1 function works, other works exactly the same with some extra checks.

Firewall logic is divided into 2 parts. Packet-wise and connection-wise. First its checked whether the connection is allowed or not, in the second part packet is checked whether it should be allowed in.

The tests also follow a similar pattern of copy-paste and making them comply. Further tests of `firewall.rs` aren't added as `firewall.rs` is already extensively tested and any change in its underlying logic still should make them pass.

To review firewall logic, which is the main change I would recommend following first the `handle_incoming_udp`. Its tcp and icmp part follow a very similar pattern and would be easier to look at.

| Process                                         | IP  | Time (ms) (Old) | Time (ms) (New) | Time Change (%)      | Throughput (Melem/s) (Old) | Throughput (Melem/s) (New) | Throughput Change (%)      |
|-------------------------------------------------|-----|-----------------|-----------------|----------------------|----------------------------|----------------------------|----------------------------|
| Inbound TCP - Rejected                          | v4  | 3.3908          | 5.7468          | **+69.5%** :red_circle: | 29.492                     | 17.401                     | **-41.0%** :red_circle:   |
| Inbound TCP - Rejected                          | v6  | 4.1948          | 6.8401          | **+63.1%** :red_circle: | 23.839                     | 14.620                     | **-38.7%** :red_circle:   |
| Inbound TCP - Accepted (Peer)                   | v4  | 1.3290          | 5.4866          | **+312.8%** :red_circle: | 75.244                     | 18.226                     | **-75.8%** :red_circle:   |
| Inbound TCP - Accepted (Peer)                   | v6  | 1.3301          | 6.6332          | **+398.7%** :red_circle: | 75.185                     | 15.076                     | **-79.9%** :red_circle:   |
| Inbound TCP - Accepted (Port)                   | v4  | 5.9637          | 8.3871          | **+40.6%** :red_circle: | 16.768                     | 11.923                     | **-28.9%** :red_circle:   |
| Inbound TCP - Accepted (Port)                   | v6  | 7.0498          | 9.7450          | **+38.2%** :red_circle: | 14.185                     | 10.262                     | **-27.6%** :red_circle:   |
| Outbound TCP - SYN                              | v4  | 4.1037          | 4.8381          | **+17.9%** :red_circle: | 24.368                     | 20.669                     | **-15.2%** :red_circle:   |
| Outbound TCP - SYN                              | v6  | 5.1511          | 5.7263          | **+11.2%** :red_circle: | 19.413                     | 17.463                     | **-10.1%** :red_circle:   |
| Outbound TCP - Other                            | v4  | 2.6330          | 2.6266          | **-0.2%** :green_circle: | 37.980                     | 38.072                     | **+0.2%** :green_circle:  |
| Outbound TCP - Other                            | v6  | 3.1599          | 3.1476          | **-0.4%** :green_circle: | 31.647                     | 31.771                     | **+0.4%** :green_circle:  |
| Outbound TCP - Accepted (Peer)                  | v4  | 1.2914          | 1.2699          | **-1.7%** :green_circle: | 77.434                     | 78.743                     | **+1.7%** :green_circle:  |
| Outbound TCP - Accepted (Peer)                  | v6  | 1.2936          | 1.3098          | **+1.2%** :red_circle:  | 77.301                     | 76.350                     | **-1.2%** :red_circle:    |
| Inbound UDP - Rejected                          | v4  | 3.3988          | 5.5694          | **+63.9%** :red_circle: | 29.422                     | 17.955                     | **-39.0%** :red_circle:   |
| Inbound UDP - Rejected                          | v6  | 4.1920          | 6.7177          | **+60.3%** :red_circle: | 23.855                     | 14.886                     | **-37.6%** :red_circle:   |
| Inbound UDP - Accepted (Peer)                   | v4  | 1.3244          | 5.4195          | **+309.3%** :red_circle: | 75.506                     | 18.452                     | **-75.6%** :red_circle:   |
| Inbound UDP - Accepted (Peer)                   | v6  | 1.3257          | 6.5999          | **+397.8%** :red_circle: | 75.430                     | 15.152                     | **-79.9%** :red_circle:   |
| Inbound UDP - Accepted (Port)                   | v4  | 5.0505          | 8.3464          | **+65.2%** :red_circle: | 19.800                     | 11.981                     | **-39.5%** :red_circle:   |
| Inbound UDP - Accepted (Port)                   | v6  | 5.8529          | 9.5536          | **+63.2%** :red_circle: | 17.086                     | 10.467                     | **-38.7%** :red_circle:   |
| Outbound UDP - Accepted (Peer)                  | v4  | 1.2832          | 1.2801          | **-0.2%** :green_circle: | 77.928                     | 78.118                     | **+0.2%** :green_circle:  |
| Outbound UDP - Accepted (Peer)                  | v6  | 1.2916          | 1.2880          | **-0.3%** :green_circle: | 77.426                     | 77.638                     | **+0.3%** :green_circle:  |
| Outbound UDP - Accepted                         | v4  | 5.6719          | 7.3268          | **+29.2%** :red_circle: | 17.631                     | 13.648                     | **-22.6%** :red_circle:   |
| Outbound UDP - Accepted                         | v6  | 6.7976          | 8.0390          | **+18.2%** :red_circle: | 14.711                     | 12.439                     | **-15.4%** :red_circle:   |



### :ballot_box_with_check: Definition of Done checklist
- [x] Commit history is clean ([requirements](../blob/main/docs/git_commit_messages_requirements.md))
- [x] README.md is updated
- [x] Functionality is covered by unit or integration tests