@lcruz99 2024-05-24
I'd ask you for next time that you can split the changes in different commits. Reviewing this PR could be made a lot easier if more information was provided: descriptive commits, nicer descriptions, that could serve as leads for me as a reviewer.