@Hasan6979 2024-05-27
> All in all, would be great to have more leads about these changes so that reviewers can have a better experience. I mean, if you take a step back, it's a really vague description (on 1 commit) for a PR with +2000 line changes on 31 files, isn't it?
I have added flow graph for the updated logic to make it easier to understand what was intended.