@Hasan6979 2024-10-21
@jjanowsk 
> TBH it feels wrong to add a feature config only to be able to run nat-lab.
So why this was decided was.
Options were make stun server as public IP.
 - problem : Sometimes public IPs are part of local networks and could potentially introduce leaks.
 Hence it was decided to stick stun ip within [RFC 1918](https://datatracker.ietf.org/doc/html/rfc1918) address limits and make private IP range configurable.