@Hasan6979 2024-11-18
> I'm still not convinced that it should be a feature, besides that it looks ok to me, +1

@stalowyjez  It is needed because atm in Apple for eg you can only allow routing, or local area connections if you allow remote access permission whereas functionally that shouldn't be the case and it should be decoupled, and it was an oversight during earlier development.