@stalowyjez 2024-11-18
@Hasan6979 that was about this local network range - IMHO this could be much cleaner and probably a bit more efficient to make it compile-time tbh